pmath:   - finished fixing #includes   - message queues now use auto-reset events instead of semaphores.   - added pmath_thread_sleep_timeout() and modified pmath_thread_send_wait() to use it. This way, we got rid of one delayed-evaluation, which always disturbs the destination thread (esp. when it is a no-op because the timeout was not reached)
richmath: mostly cosmetic changes

+++ richmath/src/graphics/config-shaper.cpp
@@ -1107,7 +1107,7 @@ void ConfigShaper::vertical_glyph_size(
   float           *ascent,
   float           *descent
 ){
-  if(info.fontinfo >= math_font_faces.length()){
+  if((int)info.fontinfo >= math_font_faces.length()){
     GlyphInfo gi;
     memcpy(&gi, &info, sizeof(GlyphInfo));
